server/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
anthropic==0.71.0
